Fix: setting pathPrefix to render images
The tutorial I used didn't have a pathPrefix in gatsby-config.js
and during the course of trouble-shooting, I put in the wrong
prefix. Now that I fixed it, it works! Committing the code now
that I have fixed the problem and deployed a working version
on the gh-pages branch with gatsby-cli.